Refactor prometheus exporter code
