+++ promqtt/promexp/__init__.py
@@ -1,0 +1,5 @@
+"""Prometheus exporter package"""
+
+from .promexp import PrometheusExporter
+
+# __all__ = ["PrometheusExporter"]
